requirements.txt
Flask==3.0.0
flask-cors
waitress==2.1.2
langchain==0.0.310
python-dotenv==1.0.0
Pillow==10.0.1
pypdfium2==4.20.0
pytesseract==0.3.10
gunicorn
openai
pypdf